Table touch is not allowed
Just like the Olympic variant of round the table, we also should not tolerate table touching anymore! We've experimented with this rule for quite a while now and we've seen more diversity in the finals and a significant reduction of the time-to-final (the two OKRs) as notorious table touchers (@mvgijssel) can't use their dirty FULL table touch smash anymore.